(#10026) Re-arrange provider and type spec files to align with Puppet.
If moved the spec files now under 'puppet' to align with how Puppet does it. This
also makes more sense as this lines up with the module/class names now.

The iptables_type_spec.rb is now just firewall_type.rb to designate it is for
the generic firewall type.

iptables_prov_spec.rb is now iptables_spec.rb. The provider part is implicit
in the path.