(GH-129) Replace errant return in autoreq block
It's not valid to use `return` within a block. We could use `next []`,
however it's probably better form to just always return the array, whether
it's populated or not. This will stop the error:

    err: Got an uncaught exception of type LocalJumpError: unexpected return

When one of the listed providers isn't selected. Which is suitable, because
this autorequire won't be suitable to any other future providers anyway.