Remove acceptance/basic_spec
This removes the legacy "basic_spec" that was used as an introduction to
module testing.  It assumes the FOSS path for the module dir.  Since the
default module dir changes in PE depending on whether or not the module
is distributed with PE or not, these basic specs have been removed from
other modules.