Update tests for EL 7
Without this patch, unit tests make no distinction between EL < 7 and
EL 7. This is a problem because the iptables save exec is different
between the two operating systems, so the tests will fail for EL 7.
This patch fixes the problem by updating the tests to include the new
exec command for EL 7.