(MODULES-1984) Perform daemon-reload on systemd
This commit patches the `manifests/linux/redhat.pp` manifest to
perform a `systemctl daemon-reload` exec on platforms that also
contain firewalld.

Prior to this commit, Specifically on Centos 7, if the systemd
deamon is not reloaded then the iptables service fails to start
because it cannot see the systemd Unit for that service.

This patch refreshes the daemon's list of Units and allows the
service to start as expected.